Only set environment variables if not already set
Currently, the run.sh script will take the environment variables from
rancher-system-agent.env, overwriting any variables that are already
set. This makes changing any environment variables that are in the env
file impossible.

After this change, only the variables that are not set in the
environment are set from the rancher-system-agent.env file.