Parse CIDR on the fly in NO_PROXY env variable
The curl command doesn't parse CIDRs in the NO_PROXY environment
variable. This causes issues when trying to contact a private IP (in the
ranche 10.0.0.0/8 for example) because the proxy likely won't be able to
contact the private IP.

The change parses the NO_PROXY variable and determines if the IP address
that curl is attempting to contact is in the CIDR of any portion of the
NO_PROXY variable. If it is in a range, then the proxy will be ignored.