Fix valid IP check
The valid_ip function did not properly exit when it detected an invalid
IP address. This caused the function to continue to try to compare a
string and an integer, which is not valid.

In addition, if NO_PROXY only contained one valid IP address, then the
for loop would continue forever. By adding '-s' the case of one valid IP
address is handled correctly.