Add support for transactional-update based systems
I've introduced support for systems that utilise transactional-update,
e.g. SLE Micro, openSUSE Leap Micro, and openSUSE MicroOS. On these
systems the filesystem is either read-only, or is snapshot based, which
means that the original install locations for CATTLE_AGENT_BIN_PREFIX
and CATTLE_AGENT_VAR_DIR are not typically writeable, depending on
whether transactional-update is being called.

This patch checks to see whether the system is using
transactional-update and if so, adapts the install locations and/or
calls transactional-update to make the appropriate changes.

This patch set supports both manually executing the script, i.e. a
Rancher "custom" installation, and initiating the deployment via
ignition/combustion (where transactional-update is called
automatically).